lsf option to stream worker stdout to master
This updates the LSFManager to allow stdout and stderr streams of workers to be streamed (optionally) to master.
The `bpeek` invocation when passed the `-f` flag behaves similar to `tail -f`. This is the default set for `bpeek_flags`.
The resultant stream is passed on to Distributed module, after some initial examination to handle the case where job startup is delayed.
The Distributed module does the parsing of initial connection information, and subsequent display of worker's stdout messages on master.

If streaming of worker's stdout messages is not desired, then the `-f` flag can be omitted from `bpeek_flags`. Then `bpeek` behaves
similar to `cat` and is used only to read the initial connection string.

As a positive side effect of this change, it avoids the earlier `redirect_stderr()` call which could have caused race conditions
with multiple async invocations of the `bpeek` method.