LSFManager optionlly stream worker stdout to mastr
This updates the LSFManager to allow stdout and stderr streams of workers to be streamed (optionally) to master. The `bpeek` invocation when passed the `-f` flag behaves similar to `tail -f`. This is the default set for `bpeek_flags`. The resultant stream is then passed on to the Distributed module after parsing the worker connection string and some initial error handling for the case where job startup is delayed or failed. The Distributed module then takes over the subsequent display of worker's stdout messages on master.

If streaming of worker's stdout messages is not desired, then the `-f` flag can be omitted from `bpeek_flags`. Then `bpeek` behaves similar to `cat` and is used only to read the initial connection string.

As a positive side effect of this change, it now also avoids the earlier `redirect_stderr()` call which could have caused race conditions with multiple async invocations of the `bpeek` method. Few methods are also renamed to have a `lsf_` prefix to avoid unnecessary clashes with other parts of the package.