Close and reopen stream after bpeek exits
Fixes #182 

Not sure is this is the right way to go about it. Rather than trying to handle a stream of empty strings we just close the stream and open a new one when bpeek terminates.

Not sure what will/should happen when stream is passed over to Distributed if user does not pass -f. Does it know what to do with empty strings or should we close the stream if the process has exited?